CM-113 Extract getTxList from actions
getTxList is not an action
